bsc#1154855 - During firstboot ayast_setup will not be executed. I've moved the active part of ayast_setup into an library which can be imported from other yast modules too.
